Moving over to index re-export pattern to ensure typings are generated
