+++ src/TsiClient.ts
@@ -1,6 +1,6 @@
 import ServerClient from "./ServerClient/ServerClient";
 import UXClient from "./UXClient/UXClient";
-import Utils from "./UXClient/Utils";
+import Utils from "./UXClient/Utils/Utils";
 
 class TsiClient {
     public server = new ServerClient();
