+++ src/UXClient/Components/AvailabilityChart/AvailabilityChart.ts
@@ -2,7 +2,7 @@ import * as d3 from "d3";
 import "./AvailabilityChart.scss";
 import LineChart from "../LineChart/LineChart";
 import DateTimePicker from "../DateTimePicker/DateTimePicker";
-import Utils from "./../../Utils";
+import Utils from "../../Utils/Utils";
 import { Component } from "./../../Interfaces/Component";
 import { ChartComponent } from "../../Interfaces/ChartComponent";
 import UXClient from "../../UXClient";
