+++ src/UXClient/Components/CategoricalPlot/CategoricalPlot.ts
@@ -1,6 +1,6 @@
 import * as d3 from 'd3';
 import { Plot } from '../../Interfaces/Plot';
-import Utils from '../../Utils';
+import Utils from '../../Utils/Utils';
 import { DataTypes } from "../../Constants/Enums";
 import { NONNUMERICTOPMARGIN, LINECHARTTOPPADDING } from "../../Constants/Constants";
