+++ src/UXClient/Components/EventsTable/EventsTable.ts
@@ -1,6 +1,6 @@
 import * as d3 from 'd3';
 import './EventsTable.scss';
-import Utils from "./../../Utils";
+import Utils from "../../Utils/Utils";
 import {Component} from "./../../Interfaces/Component";
 import { ChartComponent } from '../../Interfaces/ChartComponent';
 import { EventsTableData } from '../../Models/EventsTableData';
