+++ src/UXClient/Components/Heatmap/Heatmap.ts
@@ -1,6 +1,6 @@
 import * as d3 from 'd3';
 import './Heatmap.scss';
-import Utils from "./../../Utils";
+import Utils from "../../Utils/Utils";
 import {Component} from "./../../Interfaces/Component";
 import { Legend } from '../Legend/Legend';
 import { HeatmapCanvas} from '../HeatmapCanvas/HeatmapCanvas';
