+++ src/UXClient/Components/Legend/Legend.ts
@@ -1,6 +1,6 @@
 import * as d3 from 'd3';
 import './Legend.scss';
-import Utils from "./../../Utils";
+import Utils from "../../Utils/Utils";
 import { DataTypes, EventElementTypes } from "./../../Constants/Enums";
 import {Component} from "./../../Interfaces/Component";
 import { ChartOptions } from '../../Models/ChartOptions';
