+++ src/UXClient/Components/LineChart/LineChart.ts
@@ -1,7 +1,7 @@
 import * as d3 from 'd3';
 import { interpolatePath } from 'd3-interpolate-path';
 import './LineChart.scss';
-import Utils from "./../../Utils";
+import Utils from "../../Utils/Utils";
 import { DataTypes, YAxisStates, TooltipMeasureFormat } from "./../../Constants/Enums";
 import { LINECHARTTOPPADDING, LINECHARTCHARTMARGINS, VALUEBARHEIGHT, SERIESLABELWIDTH } from "./../../Constants/Constants";
 import {Legend} from "./../Legend/Legend";
