+++ src/UXClient/Components/ModelSearch/ModelSearch.ts
@@ -1,6 +1,6 @@
 import * as d3 from 'd3';
 import './ModelSearch.scss';
-import Utils from "./../../Utils";
+import Utils from "../../Utils/Utils";
 import {Component} from "./../../Interfaces/Component";
 import ServerClient from '../../../ServerClient/ServerClient';
 import 'awesomplete';
