+++ src/UXClient/Components/ScatterPlot/ScatterPlot.ts
@@ -5,7 +5,7 @@ import { Legend } from './../Legend/Legend';
 import { ScatterPlotData } from '../../Models/ScatterPlotData';
 import Slider from './../Slider/Slider';
 import { Tooltip } from '../Tooltip/Tooltip';
-import Utils from './../../Utils';
+import Utils from '../../Utils/Utils';
 import { TooltipMeasureFormat } from "./../../Constants/Enums";
 
 class ScatterPlot extends ChartVisualizationComponent {
