+++ src/UXClient/Components/Slider/Slider.ts
@@ -1,6 +1,6 @@
 import * as d3 from 'd3';
 import './Slider.scss';
-import Utils from "./../../Utils";
+import Utils from "../../Utils/Utils";
 import {Component} from "./../../Interfaces/Component";
 import { ChartOptions } from '../../Models/ChartOptions';
