+++ src/UXClient/Components/TimezonePicker/TimezonePicker.ts
@@ -1,7 +1,7 @@
 import * as d3 from 'd3';
 import './TimezonePicker.scss';
 import { ChartComponent } from '../../Interfaces/ChartComponent';
-import Utils from "./../../Utils";
+import Utils from "../../Utils/Utils";
 import moment from 'moment-timezone';
 
 class TimezonePicker extends ChartComponent{
