+++ src/UXClient/Interfaces/ChartVisualizationComponent.ts
@@ -1,4 +1,4 @@
-import Utils from "../Utils";
+import Utils from "../Utils/Utils";
 import {ChartComponent} from "./ChartComponent";
 import { ChartOptions } from "../Models/ChartOptions";
 import { ChartDataOptions } from "../Models/ChartDataOptions";
