+++ src/UXClient/Models/GroupedBarChartData.ts
@@ -1,4 +1,4 @@
-import Utils from "./../Utils";
+import Utils from "../Utils/Utils";
 import {ChartComponentData} from "./ChartComponentData";
 import * as d3 from 'd3';
