+++ src/UXClient/Models/TimeSeriesEvent.ts
@@ -1,4 +1,4 @@
-import Utils from "./../Utils";
+import Utils from "../Utils/Utils";
 import { TimeSeriesEventCell } from "./TimeSeriesEventCell";
 
 class TimeSeriesEvent {
