+++ src/UXClient/Models/TsqRange.ts
@@ -1,4 +1,4 @@
-import Utils from '../Utils';
+import Utils from '../Utils/Utils';
 
 class TsqRange{
     public from: Date;
