+++ src/UXClient/Utils/index.ts
@@ -1,0 +1,2 @@
+import Utils from "./Utils";
+export default Utils;
